git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/fs-py@1177691 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_diff/parse-diff.c
@@ -22,6 +22,7 @@
  */
 
 #include <stdlib.h>
+#include <stddef.h>
 #include <string.h>
 
 #include "svn_types.h"
@@ -965,8 +966,8 @@ git_start(enum parse_state *new_state, char *line, svn_patch_t *patch,
 
   while (TRUE)
     {
-      int len_old;
-      int len_new;
+      ptrdiff_t len_old;
+      ptrdiff_t len_new;
 
       new_path_marker = strstr(new_path_start, " b/");
